+++ views/householdquestionnaire.js
@@ -24,14 +24,6 @@ $('#add-shower').on('click', function () {
             );
 });
 
-var textAreaLength = 0;
-$("#openbem").on("textInput", '[key="data.household.commentary"]', function () { // I have added this event because the Commentary box in Household questionnaire was not getting saved when the assessor refreshed the page or moved to another one before losing the focus of the input (which is what triggers the onChange event)
-    if ($(this)[0].value.length - textAreaLength > 15) {
-        textAreaLength = $(this)[0].value.length;
-        $(this).trigger("change");
-    }
-});
-
 function householdquestionnaire_UpdateUI() {
     if (data.household['3b_extra_showers'] != undefined) {
         $('#more-showers').html('');